backend/requirements.txt
flask
flask-cors
httpx
beautifulsoup4
lxml
openai
python-dotenv
PyJWT
gunicorn
